requirements.txt
requests
Flask-script
Flask-PyMongo
